Add a function "get-changed-files-from-env" so we can avoid looking up the baseRef in actions
Summary:
This will help us move eslint/jest/flow-action away from needing to have the base ref checked out.

Test plan:
This is just moving some logic around. Checks should all pass.
